Major reporting and logging enhancements + Adding animated GIF support
.GifSequenceWriter: A new class dedicated to creating and configuring
the Animated GIF files.
.InvokedMethodListener: A new listener that controls logging on the
level of all invoked methods (includes special handling for first and
last methods of a test suite).
.SuiteListener: A new listener that controls logging and a fail-safe to
ensure all browsers are closed on the level of the current test suite.
.Assertions: Added support for discreet logging; adding to logs but not
as a step to the allure report.
.BrowserActions: Added support for discreet logging; adding to logs but
not as a step to the allure report.
.BrowserFactory: Added support for GIF creation, fixed issue with
attachBrowserLogs().
.ElementActions: removed deprecated method.
.ImageProcessingActions: Added support for discreet logging; adding to
logs but not as a step to the allure report.
.pom.xml: Added support for GIF creation, Added support for the new
listeners, and version bump.
.RecordManager: minor linting.
ReportManager: Added support for the new listeners, Added support for
discreet logging; adding to logs but not as a step to the allure report.
.RestActions: Added support for discreet logging; adding to logs but not
as a step to the allure report.
.ScreenshotManager: Added support for GIF creation, fixed issue with
screenshot naming convention.
.SSHActions: Added support for discreet logging; adding to logs but not
as a step to the allure report.
.Verifications: Added support for discreet logging; adding to logs but
not as a step to the allure report.
.All Test Classes were updated to remove any warnings that resulted from
the new listeners-based architecture.