905 feature request element identification name location performance boost (#935)
* internal class restructure

removed all packages named (Helper) and will make more changes to streamline usability.

Signed-off-by: MohabMohie <Mohab.MohieElDeen@outlook.com>

* the good stuff

We have a performance issue in the core of element actions. in order to do all the checks and validations that SHAFT does (especially for the type method) we have to perform multiple calls to Selenium WebDriver.
Each call comes at a significant price, especially when executing remotely via a Selenium Grid or Appium Server.
In order to minimize these calls, SHAFT will try to reuse the WebElement object, and will only try to find it again (the most costly procedure) in case it cannot reuse the existing object reference... I got this idea from TitusFortner in a chat message he sent in the Selenium TLC.
The second idea is that SHAFT will return the complete outerHTML (the element and all its components) and innerHTML of any element it attempts to find in this new class "ElementInformation" this way we will be able to perform operations on the extracted element DOM without having to call Selenium again to get the element text / textcontent / value when validating that typing was successful.
I implemented this change only for the Type method. If successfull, I should rollout to all the element actions and it should significantly enhance SHAFT's performance either by reusing existing element references (where applicable) or by extracting element information once without needing to relocate the element each time.

Signed-off-by: MohabMohie <Mohab.MohieElDeen@outlook.com>

* fix failing tests

- bug validating text was typed successfully
- sleeper bug with clipboard actions not working at all
- sleeper bug with special escaped slashes text was not being typed correctly

Signed-off-by: MohabMohie <Mohab.MohieElDeen@outlook.com>

* fix failing tests

- bug validating text was typed successfully
- sleeper bug with clipboard actions not working at all
- sleeper bug with special escaped slashes text was not being typed correctly

Signed-off-by: MohabMohie <Mohab.MohieElDeen@outlook.com>

* fix broken tests

- update default browser to be selenium naming compliant
- skip shadow dom test on firefox as it's not supported by firefox driver yet
- fix chrome 111 issue
- update google chrome searchbox locator

Signed-off-by: MohabMohie <Mohab.MohieElDeen@outlook.com>

* fixing documentation warnings

Signed-off-by: MohabMohie <Mohab.MohieElDeen@outlook.com>

* version bump for immediate release

Signed-off-by: MohabMohie <Mohab.MohieElDeen@outlook.com>

---------

Signed-off-by: MohabMohie <Mohab.MohieElDeen@outlook.com>