Added a feature to get all the rows data for a table
This feature allows users to easily get the rows data for any table that follow the standard html table format: which should have <thead> tag that has all labels in <th> tags and <tbody> which should have all rows in <tr> tags and all the data in <td> tags. for tables.
Note that this feature will only work if the html table follows this standard and does not guarantee results if any tag is missing